refactor LayoutGrid, constants
